+++ src/main/java/TheatreTicketBookingSystem/domain/SciFi.java
@@ -12,7 +12,7 @@ public class SciFi implements Movie
 {
     @Id
     @GeneratedValue
-    private int movie_id;
+    private Long movie_id;
     private int duration;
     private String title;
 
@@ -21,7 +21,7 @@ public class SciFi implements Movie
 
     }
 
-    public int getMovie_Id()
+    public Long getMovie_Id()
     {
         return movie_id;
     }
@@ -45,13 +45,13 @@ public class SciFi implements Movie
 
     public static class Builder{
 
-        private int movie_id;
+        private Long movie_id;
         private String title;
         private int duration;
 //        private Recipe recipe;
 //        private Viewer viewer;
 
-        public Builder movie_id(int movie_id) {
+        public Builder movie_id(Long movie_id) {
             this.movie_id = movie_id;
             return this;
         }
