+++ src/main/java/TheatreTicketBookingSystem/factories/ActionFactory.java
@@ -11,7 +11,7 @@ public class ActionFactory {
 
     public static Action getAction(Map<String, Object> values){
         Action factoryAction = new Action.Builder()
-                .movie_id((Integer)values.get("movie_id"))
+                .movie_id((Long)values.get("movie_id"))
                 .duration((Integer)values.get("duration"))
                 .title((String)values.get("title"))
                 .build();
