+++ src/main/java/TheatreTicketBookingSystem/factories/DocumentaryFactory.java
@@ -10,7 +10,7 @@ import java.util.Map;
 public class DocumentaryFactory {
     public static Documentary getDocumentary(Map<String, Object> values){
         Documentary factoryDocumentary = new Documentary.Builder()
-                .movie_id((Integer)values.get("movie_id"))
+                .movie_id((Long)values.get("movie_id"))
                 .duration((Integer)values.get("duration"))
                 .title((String)values.get("title"))
                 .build();
